requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain_google_genai